Extract the non-interactive auth engine into a new auth workspace crate
crates/auth-internal (package name: auth) now owns credential storage and
resolution, the blocking HTTP client, the device-flow protocol (start/poll),
backend URL resolution, org-id and federation-target resolution, the
router-config cache, and the debug-only embedded dev-CA material (dev-ca/
moves with the crate so its include_bytes! paths stay valid).

The peppy CLI keeps the interactive residue: the device-flow browser open,
verification-URL printing, and spinner move into a private run_device_flow
in the login command, and DeviceFlowOptions collapses to a bool. peppy's
error enum drops Http and NotAuthenticated (they live in auth's new
thiserror enum with verbatim Display strings) and gains a converting
AuthEngine variant; Auth(String) stays for CLI-side messages.

Engine-level integration tests move to the auth crate; command-level
login/logout/whoami tests stay in peppy.
