+++ crates/auth-internal/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "auth"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+
+[dependencies]
+config = { workspace = true }
+daemon-config = { workspace = true }
+json5-pretty = { workspace = true }
+pmi = { workspace = true, features = ["zenoh"] }
+secrecy = "0.10"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+serde_json5 = "0.2"
+thiserror = "2.0"
+tracing = "0.1"
+ureq = "3.1"
+url = "2"
+
+[dev-dependencies]
+httpmock = "0.8"
+tempfile = "3.8"
