+++ crates/auth-internal/src/client.rs
