+++ crates/auth-internal/src/discovery.rs
