+++ crates/auth-internal/src/http.rs
