+++ crates/peppy/Cargo.toml
@@ -5,6 +5,7 @@ edition.workspace = true
 autotests = false
 
 [dependencies]
+auth = { workspace = true }
 pmi = { workspace = true, features = ["zenoh", "build_zenoh"] }
 config = { workspace = true }
 daemon-config = { workspace = true }
@@ -40,10 +41,6 @@ crossterm = "0.29"
 # Terminal display width for aligning `stack list` table columns when a cell
 # (e.g. a Unicode filesystem path) contains wide or zero-width glyphs.
 unicode-width = "0.2"
-# Auth (`peppy login`): blocking HTTP for the sequential cli-config -> OIDC
-# discovery -> device-poll -> refresh -> /logout chain (matches the workspace's
-# existing ureq 3.1 in core-node-internal; no async runtime needed).
-ureq = "3.1"
 # Auto-open the browser at the device-flow verification URL on a TTY.
 open = "5"
 # Spinner shown while polling the token endpoint for device-flow approval.
